1.0.53

+++ src/main/java/com/jbnucpu/www/service/DeleteService.java
@@ -13,14 +13,23 @@ public class DeleteService {
 
     private final ContentRepository contentRepository;
 
+    private final AuthService authService;
+
     public void processNoticeDelete(Long id){
 
         noticeRepository.deleteById(id);
 
     }
 
-    public void processContentDelete(Long id){
+    public Boolean processContentDelete(Long id){
 
+        if(!authService.isAuthenticated()){
+            System.out.println("삭제 실패: 로그인 한 사용자 본인만 삭제 가능");
+            return false;
+        }
+        // TODO: 작성자만 삭제 가능
         contentRepository.deleteById(id);
+
+        return true;
     }
 }
